requirements.txt
# SafeMap - Python Dependencies
# Install with: pip install -r requirements.txt

# Core GIS libraries
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework (FastAPI backend)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Database (PostgreSQL + PostGIS)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0

# HTTP requests (for Overpass API, WFS)
requests>=2.31.0
httpx>=0.26.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization / maps (optional, for notebooks)
folium>=0.15.0
matplotlib>=3.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Code quality
ruff>=0.1.0
mypy>=1.8.0
